fix(PR #5812): port 8 new variants into renamed auto-loop-lifetime.ts + fix 4 P1 logic bugs + xrefs (12 Copilot threads)
The merge with main surfaced a rename conflict: this PR's branch had
`auto-loop-lifecycle.ts` (the pre-#5805 filename) with +1084 lines of
8-new-variant extension; main has `auto-loop-lifetime.ts` (renamed via
#5805 per `lifecycle = fixed/final` vs `lifetime = editable` convention
+ my prior P1 logic fixes). Git didn't detect the rename because both
files were substantively modified in different directions.

Resolution: ported the 8 new variants + their dispatch cases INTO
`auto-loop-lifetime.ts` (the main file), preserving #5805's P1 fixes
(refresh-substrate staleness check, brief-ack-bounded-wait →
forced-escalation transition, per-tick counter semantics). Deleted the
duplicate `auto-loop-lifecycle.ts` + `auto-loop-lifecycle.test.ts`.

Twelve Copilot threads addressed:

A. (P1 thread 1) "Per Aaron 2026-05-28" → "Per the human maintainer
   (2026-05-28)" in code + memory file headers. Also sweep across
   "Aaron names" → "the human maintainer names", "Aaron's" → "the
   human maintainer's".

B. (P2 thread 2) "7 new variants" → 8 new variants (count drift).

C. (P1 thread 3) Wildcard `*` in rule path xref
   `.claude/rules/implicit-not-explicit-in-dus-is-class-error-*.md`
   → full literal path
   `.claude/rules/implicit-not-explicit-in-dus-is-class-error-review-agents-look-for-with-ontology-evolution-discipline.md`.

D. (P1 thread 4 — ship-action unreachable post-states): ship-action
   previously transitioned directly to `tick-complete`, making the
   new `await-merge-confirmation` + `pr-loop-resolution-check` states
   UNREACHABLE per IMPLICIT-NOT-EXPLICIT rule. Changed ship-action's
   transition to `await-merge-confirmation` so the explicit post-ship
   substrate becomes reachable from the ship path.

E. (P1 thread 5 — scan-peer-prs ignores context): scan-peer-prs
   unconditionally transitioned to enter-review-mode regardless of
   whether actionable peer PRs exist. Added explicit context check:
   if `context.inflightPrs.filter(actionable).length === 0`, route to
   `free-time` (per free-time-as-valid-mode + reachability-as-offer
   invariant); else `enter-review-mode`.

F. (P1 thread 6 — free-time reachability claim drift): doc claimed
   "free-time REACHABLE-AS-OFFER from any non-terminal state" but
   only the scan-peer-prs path now routes there. Updated docblock to
   document the actual reachability paths (scan-peer-prs when
   peerActionable empty; future paths may add more) AND name Soraya
   formal-verification target explicitly. Substantive invariant
   preserved + made operationally checkable.

G. (P1 thread 7 — nextTickContext artifact-clear too broad): previous
   logic cleared `lastNamedDependency` on ANY artifact, but only
   shipped-action artifacts should clear (other artifacts like
   `verdict-only` from enter-review-mode don't ship work). Narrowed
   to `outcome.artifact.kind === "pr-opened" || "commit-pushed"`.

H. (P2 thread 8 — brief-ack docblock drift): the comment about
   counter-discipline doesn't drift now — #5805's earlier fix made
   the boundary transition through `forced-escalation` state; the
   feedback variant `CounterThresholdReached` is reserved for direct-
   dispatch callers per asymmetric-authorship. The current comment
   already reflects this state correctly.

I. (P1 thread 9 — test ship-action expectation): updated 3 failing
   tests to reflect the new routing:
   - ship-action → await-merge-confirmation (was → tick-complete)
   - operator-direction pending → await-operator-direction (was →
     brief-ack-bounded-wait conflated route)
   - runTickCycle operator-direction test expects await-operator-
     direction in transitions (was brief-ack-bounded-wait)

J. (P1 threads 10-12 — broken xrefs in memory file): replaced
   `.claude/rules/agents` → `.claude/agents/` (3 sites) since
   the agent roster lives under `.claude/agents/` not `.claude/rules/`.
   Also fixed wildcard pattern → full path for the
   IMPLICIT-NOT-EXPLICIT rule reference.

Tests: 27 pass (24 original + 3 updated for new routing semantics).

Autonomous-loop tick 2026-05-28T14:42Z resolution of PR #5812 BLOCKED
gate (12 unresolved Copilot threads + main-merge rename conflict).
